+++ modules/apps/password-policies-admin/password-policies-admin-uad-test/src/testIntegration/java/com/liferay/password/policies/admin/uad/exporter/test/PasswordPolicyUADExporterTest.java
@@ -6,7 +6,7 @@
 package com.liferay.password.policies.admin.uad.exporter.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.password.policies.admin.uad.test.PasswordPolicyUADTestUtil;
+import com.liferay.password.policies.admin.uad.test.util.PasswordPolicyUADTestUtil;
 import com.liferay.portal.kernel.model.PasswordPolicy;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
