+++ modules/apps/password-policies-admin/password-policies-admin-uad-test/src/testIntegration/java/com/liferay/password/policies/admin/uad/test/util/PasswordPolicyUADTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.password.policies.admin.uad.test;
+package com.liferay.password.policies.admin.uad.test.util;
 
 import com.liferay.portal.kernel.model.PasswordPolicy;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
